second commit

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        float regularBonus = (float) 0.3;
-        float specialBonus = (float) 0.6;
-        float totalBonus = regularBonus + specialBonus;
+        double regularBonus =  0.3;
+        double specialBonus =  0.6;
+        double totalBonus = regularBonus + specialBonus;
         System.out.println(totalBonus);
     }
 }
